Separated gui into gui.view and gui.controller and moved the classes accordingly
